refactoring attribute setting, and fix css for front
